requirements.txt
flask
flask-smorest
gunicorn
pytesseract
python-dotenv